Move action-buttons component to it's own file
